Rename enhanced_interactive_poc.py to xslt_mapping_extractor_poc.py
## Rename Rationale
- **Descriptive naming**: Clearly conveys purpose as XSLT mapping extraction tool
- **Professional**: Production-ready naming convention vs "enhanced_interactive_poc"
- **Self-documenting**: Name immediately indicates functionality
- **Achievement-focused**: Emphasizes the mapping extraction breakthrough (85.7% coverage)

## Files Updated
- **Main file**: enhanced_interactive_poc.py → xslt_mapping_extractor_poc.py
- **Documentation**: POC_IMPROVEMENT_PLAN.md reference updated
- **Analysis tool**: analyze_poc_results.py description updated
- **Test file**: test_enhanced_poc.py import statement updated
- **Utility script**: clear_poc_results.sh usage instruction updated

## Preserved References
- **Historical data**: poc_results/ files maintain original references for accuracy
- **Archive docs**: Archived documentation preserves historical context

The new name better reflects the tool's role as a production-quality XSLT business mapping extractor that achieved the breakthrough semantic chunking results.
